refactor: move files from src/pages that are not pages
