+++ src/interfaces/ITextParams.tsx
