+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import './globals.css';
+import '../globals.css';
 import type { AppProps } from 'next/app';
 import { ToastContainer } from 'react-toastify';
 import { SessionProvider } from 'next-auth/react';
