+++ src/pages/sign-up.tsx
@@ -9,8 +9,8 @@ import signUp from '@/src/api/signUp';
 import createCustomerDraft from '@/src/helpers/commercetools/customerDraft';
 import ErrorMessage from '@/src/components/ErrorMessage';
 import NamesClients from '@/src/helpers/commercetools/consts';
+import { IFormInput } from '@/src/interfaces/IFormInput';
 import InputEmail from '../components/InputEmail';
-import { IFormInput } from './interfaces/IFormInput';
 import InputPassword from '../components/InputPassword';
 import InputFirstName from '../components/InputFirstName';
 import InputLastName from '../components/InputLastName';
